+++ virtusize/src/main/java/com/virtusize/android/data/local/VirtusizeEnvironment.kt
@@ -4,6 +4,7 @@ package com.virtusize.android.data.local
  * This enum contains all available Virtusize environments
  */
 enum class VirtusizeEnvironment {
+    TESTING,
     STAGING,
     GLOBAL,
     JAPAN,
@@ -16,7 +17,8 @@ enum class VirtusizeEnvironment {
  */
 fun VirtusizeEnvironment.defaultApiUrl(): String {
     return when (this) {
-        VirtusizeEnvironment.STAGING -> "https://staging.virtusize.jp"
+        VirtusizeEnvironment.TESTING -> "https://testing.virtusize.jp"
+        VirtusizeEnvironment.STAGING -> "https://staging.virtusize.com"
         VirtusizeEnvironment.GLOBAL -> "https://api.virtusize.com"
         VirtusizeEnvironment.JAPAN -> "https://api.virtusize.jp"
         VirtusizeEnvironment.KOREA -> "https://api.virtusize.kr"
@@ -29,9 +31,10 @@ fun VirtusizeEnvironment.defaultApiUrl(): String {
  */
 fun VirtusizeEnvironment.eventApiUrl(): String {
     return when (this) {
-        VirtusizeEnvironment.STAGING -> "https://events.staging.virtusize.jp"
-        VirtusizeEnvironment.JAPAN -> "https://events.virtusize.jp"
+        VirtusizeEnvironment.TESTING -> "https://events.testing.virtusize.jp"
+        VirtusizeEnvironment.STAGING -> "https://events.staging.virtusize.com"
         VirtusizeEnvironment.GLOBAL -> "https://events.virtusize.com"
+        VirtusizeEnvironment.JAPAN -> "https://events.virtusize.jp"
         VirtusizeEnvironment.KOREA -> "https://events.virtusize.kr"
     }
 }
@@ -42,7 +45,8 @@ fun VirtusizeEnvironment.eventApiUrl(): String {
  */
 fun VirtusizeEnvironment.servicesApiUrl(): String {
     return when (this) {
-        VirtusizeEnvironment.STAGING -> "https://services.virtusize.jp/stg"
+        VirtusizeEnvironment.TESTING -> "https://services.virtusize.jp/stg"
+        VirtusizeEnvironment.STAGING -> "https://services.virtusize.com/stg"
         VirtusizeEnvironment.GLOBAL -> "https://services.virtusize.com"
         VirtusizeEnvironment.JAPAN -> "https://services.virtusize.jp"
         VirtusizeEnvironment.KOREA -> "https://services.virtusize.kr"
@@ -56,8 +60,9 @@ fun VirtusizeEnvironment.servicesApiUrl(): String {
 fun VirtusizeEnvironment.virtusizeUrl(): String {
     return when (this) {
         VirtusizeEnvironment.STAGING,
-        VirtusizeEnvironment.JAPAN -> "https://static.api.virtusize.jp"
         VirtusizeEnvironment.GLOBAL -> "https://static.api.virtusize.com"
+        VirtusizeEnvironment.TESTING,
+        VirtusizeEnvironment.JAPAN -> "https://static.api.virtusize.jp"
         VirtusizeEnvironment.KOREA -> "https://static.api.virtusize.kr"
     }
 }
@@ -70,8 +75,8 @@ fun VirtusizeEnvironment.virtusizeUrl(): String {
  */
 fun VirtusizeEnvironment.virtusizeRegion(): VirtusizeRegion {
     return when (this) {
-        VirtusizeEnvironment.STAGING, VirtusizeEnvironment.JAPAN -> VirtusizeRegion.JP
-        VirtusizeEnvironment.GLOBAL -> VirtusizeRegion.COM
+        VirtusizeEnvironment.STAGING, VirtusizeEnvironment.GLOBAL -> VirtusizeRegion.COM
+        VirtusizeEnvironment.TESTING, VirtusizeEnvironment.JAPAN -> VirtusizeRegion.JP
         VirtusizeEnvironment.KOREA -> VirtusizeRegion.KR
     }
 }
@@ -82,7 +87,7 @@ fun VirtusizeEnvironment.virtusizeRegion(): VirtusizeRegion {
  */
 fun VirtusizeEnvironment.virtusizeWebViewEnv(): String {
     return when (this) {
-        VirtusizeEnvironment.STAGING -> "staging"
+        VirtusizeEnvironment.TESTING, VirtusizeEnvironment.STAGING -> "staging"
         else -> "production"
     }
 }