+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.4.1
+
+* Use the staging API URLs for staging
+* Add the `testing` env for internal use
+
 ## 2.4.0
 
 * Add ktlint